gui decomposition starting;
