+++ CMakeLists.txt
@@ -34,6 +34,8 @@ set(OPENTOMB_SRCS
     src/core/vmath.h
     src/gui/gui.cpp
     src/gui/gui.h
+    src/gui/gui_inventory.cpp
+    src/gui/gui_inventory.h
     src/fmv/tiny_codec.h
     src/fmv/tiny_codec.c
     src/fmv/stream_codec.h
