+++ autoexec.lua
@@ -17,7 +17,8 @@ mlook(1);
 freelook(0);
 cam_distance(1024.0);
 noclip(0);
-playVideo(base_path .. "data/tr3/fmv/logo.rpl");
+--playVideo(base_path .. "data/tr3/fmv/logo.rpl");
+playVideo(base_path .. "data/tr3/fmv/Crsh_Eng.rpl");
 --setgamef(1, 17)
 --loadMap(base_path .. "tests/altroom3/LEVEL1.PHD");
 --loadMap(base_path .. "tests/heavy1/LEVEL1.PHD");
